requirements.txt
 torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
pyyaml>=6.0
tqdm>=4.66.0
pycocotools>=2.0.7  # For COCO-style annotations if needed
